Create TimeInterface
This is an interface to show how the three cost classes should be implemented.  Time, cost and effort interfaces are required so for each concrete type of unit we are required to implement conversions to and from a base unit of our choice.  Calculations will be done in this base unit, then returned to concrete form to be displayed.